Refactored keycode field - extracted `KeyCodeField` method to it's own script, folder, and class - updated use of migrated method
