+++ Editor/PropertyDrawers/GUIInputFields.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1bf02816e89e4374abd49b3aff0a5ab5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
