+++ Editor/PropertyDrawers/GUIInputFields/KeyCodeField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0d529b6e2afea124986ebec7446fae37
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
